@@ -2,7 +2,6 @@ import functools
 import inspect
 import operator
 import os
-import posixpath
 from urllib.parse import quote, urljoin, urlparse, urlunparse
 import warnings
 from webob.acceptparse import Accept
@@ -2338,14 +2337,7 @@ class StaticURLInfo:
             pathspec = f'{pkg_name}:{pkg_subpath}{subpath}'
             overrides = registry.queryUtility(IPackageOverrides, name=pkg_name)
             if overrides is not None:
-                resource_name = posixpath.join(pkg_subpath, subpath)
-                sources = overrides.filtered_sources(resource_name)
-                for source, filtered_path in sources:
-                    rawspec = source.get_path(filtered_path)
-                    if hasattr(source, 'pkg_name'):
-                        rawspec = f'{source.pkg_name}:{rawspec}'
-                    break
-
+                rawspec = overrides.get_spec(f'{pkg_subpath}{subpath}')
         else:
             pathspec = pkg_subpath + subpath
 
